+++ http/param/param_test.go
@@ -451,6 +451,72 @@ func TestParser_Parse_PathParam_FuncNotDefinedError(t *testing.T) {
 	assert.Error(t, err)
 }
 
+type structWithFormParams struct {
+	Subject string            `param:"form=subject"`
+	Amount  *int              `param:"form=amount"`
+	Object  *maybeShinyObject `param:"form=object"`
+	Nothing string            `param:"form=nothing"`
+}
+
+func TestParser_Parse_FormParam(t *testing.T) {
+	r := chi.NewRouter()
+	p := DefaultParser()
+	result := structWithFormParams{
+		Nothing: "should be replaced",
+	}
+	expected := structWithFormParams{
+		Subject: "world",
+		Amount:  ptr(69),
+		Object: &maybeShinyObject{
+			IsShiny: true,
+			Object:  "apples",
+		},
+		Nothing: "",
+	}
+	var parseError error
+	r.Post("/hello/objects", func(_ http.ResponseWriter, r *http.Request) {
+		parseError = p.Parse(r, &result)
+	})
+
+	urlEncodedBodyContent := "subject=world&amount=69&object=shiny-apples&nothing="
+	urlEncodedBody := strings.NewReader(urlEncodedBodyContent)
+	req := httptest.NewRequest(http.MethodPost, "https://test.com/hello/objects", urlEncodedBody)
+	req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
+	r.ServeHTTP(httptest.NewRecorder(), req)
+
+	require.NoError(t, parseError)
+	assert.Equal(t, expected, result)
+}
+
+func TestParser_Parse_FormParam_NoParams(t *testing.T) {
+	p := DefaultParser()
+	result := structWithFormParams{
+		Subject: "should be replaced",
+		Amount:  ptr(123), // should be zeroed out
+		Nothing: "should be replaced",
+	}
+	expected := structWithFormParams{
+		Subject: "",
+		Amount:  nil,
+		Object:  nil,
+		Nothing: "",
+	}
+	var parseError error
+
+	r := chi.NewRouter()
+	r.Post("/hello/objects", func(_ http.ResponseWriter, r *http.Request) {
+		parseError = p.Parse(r, &result)
+	})
+
+	// Empty form body
+	req := httptest.NewRequest(http.MethodPost, "https://test.com/hello/objects", strings.NewReader(""))
+	req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
+	r.ServeHTTP(httptest.NewRecorder(), req)
+
+	require.NoError(t, parseError)
+	assert.Equal(t, expected, result)
+}
+
 type otherFieldsStruct struct {
 	Q     string `param:"query=q"`
 	Other string `json:"other"`